refactor: lotto 패키지 기준으로 패키지 구분
