--- src/main/java/Application.java
@@ -1,9 +1,0 @@
-import controller.LottoController;
-
-public class Application {
-
-    public static void main(String[] args) {
-        LottoController lottoController = new LottoController();
-        lottoController.startLotto();
-    }
-}
