--- src/main/java/domain/lotto/LottoResult.java
@@ -1,24 +1,0 @@
-package domain.lotto;
-
-import domain.lotto.prize.LottoPrize;
-
-import java.util.Map;
-
-public class LottoResult {
-
-    private final Map<LottoPrize, Integer> prizeCounts;
-    private final double earningRate;
-
-    public LottoResult(Map<LottoPrize, Integer> prizeCounts, double earningRate) {
-        this.prizeCounts = prizeCounts;
-        this.earningRate = earningRate;
-    }
-
-    public Map<LottoPrize, Integer> getPrizeCounts() {
-        return prizeCounts;
-    }
-
-    public double getEarningRate() {
-        return earningRate;
-    }
-}
