--- src/main/java/domain/lotto/LottoResultCalculator.java
@@ -1,51 +1,0 @@
-package domain.lotto;
-
-import domain.lotto.prize.LottoPrize;
-import domain.lotto.store.LottoStore;
-import domain.lotto.ticket.LottoTicket;
-
-import java.util.*;
-
-public class LottoResultCalculator {
-
-    private final List<LottoTicket> lottoTickets;
-    private final LottoWinningNumber lottoWinningNumber;
-
-    public LottoResultCalculator(List<LottoTicket> lottoTickets, LottoWinningNumber lottoWinningNumber) {
-        this.lottoTickets = lottoTickets;
-        this.lottoWinningNumber = lottoWinningNumber;
-    }
-
-    public LottoResult calculateLottoResult() {
-        double earningRate = calculateEarningRate();
-        Map<LottoPrize, Integer> prizeCounts = countLottoPrize();
-
-        return new LottoResult(prizeCounts, earningRate);
-    }
-
-    private Map<LottoPrize, Integer> countLottoPrize() {
-        Map<LottoPrize, Integer> prizeCounts = new HashMap<>();
-        Arrays.stream(LottoPrize.values())
-                .forEach(prize -> prizeCounts.put(prize, 0));
-
-        lottoTickets.stream()
-                .map(lottoTicket -> LottoPrize.findPrize(lottoTicket, lottoWinningNumber))
-                .filter(Optional::isPresent)
-                .map(Optional::get)
-                .forEach(lottoPrize -> prizeCounts.computeIfPresent(lottoPrize, (key, value) -> value + 1));
-
-        return prizeCounts;
-    }
-
-    private double calculateEarningRate() {
-        double lottoCost = lottoTickets.size() * LottoStore.LOTTO_PRICE;
-        int prizeMoneySum = lottoTickets.stream()
-                .map(lottoTicket -> LottoPrize.findPrize(lottoTicket, lottoWinningNumber))
-                .filter(Optional::isPresent)
-                .map(Optional::get)
-                .mapToInt(LottoPrize::getPrizeMoney)
-                .sum();
-
-        return prizeMoneySum / lottoCost;
-    }
-}
