--- src/main/java/domain/lotto/ticket/generator/LottoTicketGenerator.java
@@ -1,9 +1,0 @@
-package domain.lotto.ticket.generator;
-
-import domain.lotto.LottoNumber;
-
-import java.util.List;
-
-public interface LottoTicketGenerator {
-    List<LottoNumber> generate();
-}
