+++ src/main/java/lotto/domain/lotto/prize/LottoPrize.java
@@ -1,0 +1,38 @@
+package lotto.domain.lotto.prize;
+
+import lotto.domain.lotto.LottoWinningNumber;
+import lotto.domain.lotto.prize.judgment.LottoPrizeJudgment;
+import lotto.domain.lotto.prize.judgment.impl.*;
+import lotto.domain.lotto.ticket.LottoTicket;
+
+import java.util.Arrays;
+import java.util.Optional;
+
+public enum LottoPrize {
+
+    FIRST_PRIZE(2_000_000_000, new FirstLottoPrizeJudgment()),
+    SECOND_PRIZE(30_000_000, new SecondLottoPrizeJudgment()),
+    THIRD_PRIZE(1_500_000, new ThirdLottoPrizeJudgment()),
+    FOURTH_PRIZE(50_000, new FourthLottoPrizeJudgment()),
+    FIFTH_PRIZE(5_000, new FifthLottoPrizeJudgment());
+
+    LottoPrize(int prizeMoney, LottoPrizeJudgment prizeJudgment) {
+        this.prizeMoney = prizeMoney;
+        this.prizeJudgment = prizeJudgment;
+    }
+
+    private final int prizeMoney;
+
+    private final LottoPrizeJudgment prizeJudgment;
+
+    public int getPrizeMoney() {
+        return prizeMoney;
+    }
+
+    public static Optional<LottoPrize> findPrize(LottoTicket lottoTicket, LottoWinningNumber lottoWinningNumber) {
+        return Arrays.stream(values())
+                .filter(e -> e.prizeJudgment.judge(lottoTicket, lottoWinningNumber))
+                .findFirst();
+    }
+
+}
