+++ src/main/java/lotto/domain/lotto/prize/judgment/LottoPrizeJudgment.java
@@ -1,0 +1,9 @@
+package lotto.domain.lotto.prize.judgment;
+
+import lotto.domain.lotto.LottoWinningNumber;
+import lotto.domain.lotto.ticket.LottoTicket;
+
+public interface LottoPrizeJudgment {
+
+    boolean judge(LottoTicket lottoTicket, LottoWinningNumber lottoWinningNumber);
+}
