+++ src/main/java/lotto/domain/lotto/prize/judgment/impl/FifthLottoPrizeJudgment.java
@@ -1,0 +1,20 @@
+package lotto.domain.lotto.prize.judgment.impl;
+
+import lotto.domain.lotto.LottoNumber;
+import lotto.domain.lotto.LottoWinningNumber;
+import lotto.domain.lotto.prize.judgment.LottoPrizeJudgment;
+import lotto.domain.lotto.ticket.LottoTicket;
+import util.ListUtils;
+
+import java.util.List;
+
+public class FifthLottoPrizeJudgment implements LottoPrizeJudgment {
+
+    @Override
+    public boolean judge(LottoTicket lottoTicket, LottoWinningNumber lottoWinningNumber) {
+        List<LottoNumber> lottoNumbers = lottoTicket.getLottoNumbers();
+        List<LottoNumber> winningNumbers = lottoWinningNumber.getLottoNumbers();
+
+        return ListUtils.countCommonElements(lottoNumbers, winningNumbers) == 3;
+    }
+}
