+++ src/main/java/lotto/domain/lotto/ticket/generator/LottoTicketGenerator.java
@@ -1,0 +1,9 @@
+package lotto.domain.lotto.ticket.generator;
+
+import lotto.domain.lotto.LottoNumber;
+
+import java.util.List;
+
+public interface LottoTicketGenerator {
+    List<LottoNumber> generate();
+}
