+++ src/test/java/domain/lotto/LottoNumberTest.java
@@ -1,5 +1,6 @@
 package domain.lotto;
 
+import lotto.domain.lotto.LottoNumber;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
